+++ WormBrowser/OWMetaDataViewController.m
@@ -215,7 +215,7 @@
 
 -(void) handleButtonTap:(id)sender
 {
-    int tag = [sender tag];
+    NSInteger tag = [sender tag];
     
     switch (tag) {
         case 0:
